requirements.txt
absl-py==2.3.1
asgiref==3.9.1
attrs==25.3.0
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
contourpy==1.3.3
cycler==0.12.1
django==5.2.5
flatbuffers==25.2.10
fonttools==4.59.0
jax==0.7.0
jaxlib==0.7.0
kiwisolver==1.4.9
matplotlib==3.10.5
mediapipe==0.10.21
ml-dtypes==0.5.3
msgpack==1.1.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.4.0
packaging==25.0
pgvector==0.3.6
pillow==11.3.0
pip==25.2
protobuf==4.25.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
ruff==0.12.8
scipy==1.16.1
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
sqlparse==0.5.3
